links are just normal external references, with a sugared syntax.
when you have a @<recordid> value in a row, it is a link.
you can't do anything in that row, but you can refer to it from other rows
using the `link` function, so if "x" is "@riwue2", on "y" you can do
`=.x | link | .attr` to get "attr" from the record @riwue2.

this is the same as doing `=$riwue2.attr` (exactly the same, the `link`
function just proxies you to the $ variable), but it is good if you're
using default fields with formulas (not implemented yet) and you want
to keep the formula, just replace the link, for example.

also, linked rows have the search/combobox/select thing that helps in
selecting other records to link.

maybe linked arrays will be implemented in the future.