fix reference gathering for links.
stupid coding mistakes were preventing references to rows that
represent linked records to be gathered as global references to
the linked records and thus have their update triggered when the
given record is changed.

now this is fixed, but still remains the question of when will
these references be gathered at the application start.