change the record loading model to use the .changes listener.
this is good for environments where PouchDB replications are happening.

it may be good for initial load performance also, as we've removed the need
for sending the records to Elm and then passing them back so they can be
calculated.

also, we don't call `changedValue` for each row of each record, only for the
first row of each record, since that single call will trigger updates on each
other rows if they are formulas.

actually, we don't really call `changedValue`, we've replaced that with two
more narrowly scoped functions, so we can batch them better, but it is the same.

also, the code is simpler, I think.